+++ app/opulence/orm/IEntityRegistry.php
@@ -10,7 +10,6 @@ interface IEntityRegistry
 {
     /**
      * Clears all the contents of the registry
-     * All entities will appear
      * This should only be called through a unit of work
      */
     public function clear();
@@ -48,15 +47,6 @@ interface IEntityRegistry
     public function getEntity($className, $id);
 
     /**
-     * Gets the entity Id
-     *
-     * @param object $entity The entity whose Id we want
-     * @return mixed The Id
-     * @throws ORMException Thrown if no Id getter has been registered for this entity
-     */
-    public function getEntityId($entity);
-
-    /**
      * Gets the entity state for the input entity
      *
      * @param object $entity The entity to check
@@ -73,15 +63,6 @@ interface IEntityRegistry
     public function getObjectHashId($object);
 
     /**
-     * Checks whether or not an object has changed
-     *
-     * @param object $entity The entity to check
-     * @return bool True if the entity has changed, otherwise false
-     * @throws ORMException Thrown if the entity was not registered
-     */
-    public function hasChanged($entity);
-
-    /**
      * Gets whether or not an entity is registered
      *
      * @param object $entity The entity to check
@@ -90,40 +71,12 @@ interface IEntityRegistry
     public function isRegistered($entity);
 
     /**
-     * Registers a comparison function for a class, which speeds up the check for updates
-     * Registering a comparison function for a class will overwrite any previously-set comparison functions for that class
-     *
-     * @param string $className The name of the class whose comparison function we're registering
-     * @param callable $function The function that takes two instances of the same class and returns whether or not
-     *      they're considered identical
-     */
-    public function registerComparisonFunction($className, callable $function);
-
-    /**
      * Registers an entity
      *
      * @param object $entity The entity to register
      * @throws ORMException Thrown if there was an error registering the entity
      */
     public function registerEntity(&$entity);
-
-    /**
-     * Registers functions that get an Id and set the Id for all instances of the input class name
-     *
-     * @param string $className The name of the class whose Id getter function we're registering
-     * @param callable $getter The function that accepts an entity as a parameter and returns its Id
-     * @param callable $setter The function that accepts an entity and new Id as parameters and sets the Id
-     */
-    public function registerIdAccessors($className, callable $getter, callable $setter = null);
-
-    /**
-     * Sets the entity Id
-     *
-     * @param object $entity The entity whose Id we're setting
-     * @param mixed $id The Id to set
-     * @throws ORMException Thrown if no Id setter has been registered for this entity
-     */
-    public function setEntityId($entity, $id);
 
     /**
      * Sets an entity's state